added calender event

+++ src/main/java/com/example/creation/dto/response/SemesterResponseDto.java
@@ -5,7 +5,7 @@ public record SemesterResponseDto(
 
         String semesterName,
 
-        Long facultyId,
+        int facultyId,
 
         String facultyName
 ) {
